feat: implement Phase 2 of ROADMAP.md
- Migrate global/singleton lyrics sync state, parsing, and playback timing to useLyrics composable.
- Extract lyric modal editor and item alteration actions to useLyricEditor composable.
- Clean up and streamline App.vue to use the new composables.
- Manage AnimationManager and scroll transitions centrally in useLyrics.
